The most significant changes involve the addition of new methods to the `IStronglyTypedId` interface and the `OneOfExtensions` class.
In the `IStronglyTypedId` interface, a new `ToString()` method has been added. This method is expected to be used for converting the `Guid Value` to a string representation.

In the `OneOfExtensions` class, two new methods have been added: `MatchAsync<T0, T1, TResult>` and `MatchAsync<T0, T1, T2, TResult>`. These methods are extension methods for `Task<OneOf<T0, T1>>` and `Task<OneOf<T0, T1, T2>>` objects respectively. They take functions as parameters, await the task, call the `Match` method on the result, pass in the functions, await the result of the `Match` method, and then return it.

List of changes:

1. Added a `ToString()` method to the `IStronglyTypedId` interface for converting the `Guid Value` to a string representation.
2. Added a `MatchAsync<T0, T1, TResult>` method to the `OneOfExtensions` class. This method is an extension method for a `Task<OneOf<T0, T1>>` object and takes two functions as parameters.
3. Added a `MatchAsync<T0, T1, T2, TResult>` method to the `OneOfExtensions` class. This method is an extension method for a `Task<OneOf<T0, T1, T2>>` object and takes three functions as parameters.